Patch : Quiz start 실행시 반환값 제시어 바꿈

+++ mung-BE/src/main/java/com/mung/mung/db/entity/Game.java
@@ -34,7 +34,7 @@ public class Game {
 
     private LocalDateTime endTime;
 
-    private int ranQuiz;
+    private Long ranQuiz;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "room_id")
